% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{height_data}
\alias{height_data}
\title{Howell !Kung Demography Data}
\format{
\subsection{\code{height_data}}{

A data frame containing demographic measurements of !Kung San individuals:
\describe{
\item{height}{Height in cm}
\item{weight}{Weight in kg}
\item{age}{Age in years}
\item{male}{Gender indicator}
\item{age.at.death}{Age at death in years, if deceased}
\item{alive}{Indicator of whether the individual is still alive}
}
}
}
\source{
Downloaded from the University of Toronto TSpace repository:
\url{https://tspace.library.utoronto.ca/handle/1807/10395}
}
\usage{
height_data
}
\description{
Demographic data from Kalahari !Kung San people collected by Nancy Howell.
}
\references{
Howell, N. Demographic data from the Kalahari !Kung San people.
}
\keyword{datasets}
