% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{medisr_install_packages}
\alias{medisr_install_packages}
\title{Install all packages used for the Medical Statistics course}
\usage{
medisr_install_packages()
}
\description{
A simple high-level function to install all packages needed.
}
\examples{
\dontrun{
medisr_install_packages()
}
}
