% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prostate_cancer}
\alias{prostate_cancer}
\alias{prostate_cancer_train}
\alias{prostate_cancer_test}
\title{Prostate Cancer Data}
\format{
The prostate_cancer dataset contains two data frames,
\code{prostate_cancer_train} and \code{prostate_cancer_test}, which can be loaded
together using \code{data(prostate_cancer)}.

Both data frames contain the following variables:
\describe{
\item{cancer}{Prostate cancer confirmed by biopsy
(1 = yes, 0 = no).}
\item{symptomatic}{Indicator of whether the patient has trouble urinating
and/or decreased force of urine stream.}
\item{biomarkerk}{Concentration of an unspecified biomarker extracted
from a blood sample, where k denotes the biomarker number 1-12.}
}

\code{prostate_cancer_train} is the training set and \code{prostate_cancer_test} is
the test set.
}
\source{
Decencière, E., Cazuguel, G., Zhang, X., Thibault, G., Quellec, G.,
Lamard, M., Conze, P.-H., Massin, P., Cochener, B., Trone, C., Gain, P.,
Ordonez, R., Abràmoff, M. D., and others. (2014). Feedback on a publicly
distributed database: the Messidor database. Image Analysis &
Stereology, 33(3), 231--234.

Abràmoff, M. D., Folk, J. C., Han, D. P., Walker, J. D., Williams, D. F.,
Russell, S. R., Massin, P., Cochener, B., Gain, P., Tang, L., Lamard, M.,
Moga, D. C., Quellec, G., and Niemeijer, M. (2013). Automated analysis of
retinal images for detection of referable diabetic retinopathy.
JAMA Ophthalmology, 131(3), 351--357.
}
\description{
Data for predicting prostate cancer from blood sample measurements.
}
\details{
The population consists of patients screened for prostate cancer by biopsy.
The data are sham data constructed by replicating the structure of the
Messidor-2 dataset. Variables have been renamed and the data have been
split into a training set and a test set.
}
\references{
Decencière, E., Cazuguel, G., Zhang, X., Thibault, G., Quellec, G.,
Lamard, M., Conze, P.-H., Massin, P., Cochener, B., Trone, C., Gain, P.,
Ordonez, R., Abràmoff, M. D., and others. (2014). Feedback on a publicly
distributed database: the Messidor database. Image Analysis &
Stereology,
33(3), 231--234.

Abràmoff, M. D., Folk, J. C., Han, D. P., Walker, J. D., Williams, D. F.,
Russell, S. R., Massin, P., Cochener, B., Gain, P., Tang, L., Lamard, M.,
Moga, D. C., Quellec, G., and Niemeijer, M. (2013). Automated analysis of
retinal images for detection of referable diabetic retinopathy.
JAMA Ophthalmology, 131(3), 351--357.
}
