% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SUA_data}
\alias{SUA_data}
\title{Serum Uric Acid and Cardiovascular Disease Risk Factors}
\format{
\subsection{\code{SUA_data}}{

A data frame with 6,119 rows and 14 columns:
\describe{
\item{ID}{Unique identifier for each participant}
\item{Age}{Age of the participant in years}
\item{Sex}{Sex of the participant. 1 = male; 2 = female}
\item{BMI}{Body mass index of the participant}
\item{SBP}{Systolic blood pressure of the participant}
\item{DBP}{Diastolic blood pressure of the participant}
\item{FBG}{Fasting blood glucose level of the participant}
\item{TC}{Total cholesterol level of the participant}
\item{Cr}{Serum creatinine level of the participant}
\item{GFR}{Glomerular filtration rate of the participant}
\item{UA}{Serum uric acid level in the participant's blood}
\item{Times}{Number of medical follow-up visits for the participant}
\item{Hypertension}{Hypertension status. 1 = no; 2 = yes}
\item{Hyperglycemia}{Hyperglycemia status. 1 = no; 2 = yes}
\item{Dyslipidemia}{Dyslipidemia status. 1 = no; 2 = yes}
}
}
}
\source{
Luo, Y., Wu, Q., Meng, R., Lian, F., Jiang, C., Hu, M., Wang, Y.,
& Ma, H. (2023). Associations of serum uric acid with cardiovascular
disease risk factors: a retrospective cohort study in Southeastern China
\link{Data set}. Zenodo.
https://doi.org/10.5061/dryad.z08kprrk1

Zenodo record:
https://zenodo.org/records/8292712

Data file:
https://zenodo.org/records/8292712/files/SUA_CVDs_risk_factors.csv
}
\usage{
SUA_data
}
\description{
This dataset contains serum uric acid measurements, cardiovascular disease
risk factors, and basic demographic information for participants in a
retrospective cohort study conducted every two years from 2010 to 2018 in
Hangzhou, Zhejiang Province, Southeastern China. A total of 6,119
participants aged 40 years and above who underwent at least three physical
examinations were enrolled.
}
\keyword{datasets}
