% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medisr-package.R
\docType{package}
\name{medisr-package}
\alias{medisr}
\alias{medisr-package}
\title{medisr: A toolbox for Medical Statistics using R}
\description{
This package provides tools we use for the course Medical Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanMazJen/medisr}
  \item \url{https://danmazjen.github.io/medisr/}
  \item Report bugs at \url{https://github.com/DanMazJen/medisr/issues}
}

}
\author{
\strong{Maintainer}: Daniel Mazhari-Jensen \email{daskmaje@hst.aau.dk}

Authors:
\itemize{
  \item Daniel Mazhari-Jensen \email{daskmaje@hst.aau.dk}
}

}
\keyword{internal}
